# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beauty Parlor API",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/appointments": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "View Appointment",
                "description": "Appointment user and invalidate token",
                "operationId": "Appointment",
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Create Appointment",
                "description": "Appointment user and invalidate token",
                "operationId": "Create Appointment",
                "requestBody": {
                    "description": "fill the",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "time"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-05-01"
                                    },
                                    "time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "11:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/appointment-history": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "View Appointment History",
                "description": "Appointment user and invalidate token",
                "operationId": "View Appointment History",
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/appointment/{id}": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "View Appointment by id",
                "description": "Appointment user and invalidate token",
                "operationId": "View Appointment by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Appointment to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Appointment"
                ],
                "summary": "update appointment status",
                "description": "Appointment user and invalidate token",
                "operationId": "update appointment status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Appointment history to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "Approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/appointment-history/{id}": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "View Appointment History by id",
                "description": "Appointment user and invalidate token",
                "operationId": "View Appointment History by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Appointment to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/appointment-user-history/{id}": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "View Appointment History by user id",
                "description": "Appointment user and invalidate token",
                "operationId": "View Appointment History by user id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Appointment history to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}